// app/src/app/Models/AduanasViewModel.ts
export class Aduana {
    Id: number;
    Aduana: string;
    Estado: string;
    esta_Id: number;
    Ciudad: string;
    ciud_Id: number;
    adua_Creacion?: number;
    adua_Modifica?: number;
    adua_FechaCreacion?: string;
    adua_FechaModifica?: string;
    Creacion?: string;
    Modifica?: string;
}